+++ src/app/core_api.h
@@ -16,6 +16,7 @@
 #undef OUTPUT
 #include "cc_core.h"
 #include "cc_dht.h"
+#include "cc_ledstrip.h"
 #pragma pop_macro("OUTPUT")
 #pragma pop_macro("DISABLED")
 #pragma pop_macro("LOW")